Complete rewamp of Weld support
This approach uses the annotations from metric-annotation, and annotated methods
are then dynamically annotated with it's Weld counterpart in order to wire
up the interceptors.

All tests pass and the only changes are that ExceptionMeteredTest has been
altered to cater for the fact that this module (currently) doesn't reset
the MetricsRegistry between each run.